third_party/blink/web_tests/platform/generic/css3/blending/blend-mode-property-parsing-invalid-expected.txt

Test the parsing of the mix-blend-mode property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Too many parameters : overlay overlay
PASS cssRule.type is 1
PASS declaration.length is 0
PASS declaration.getPropertyValue('mix-blend-mode') is ""


Wrong type : "5px"
PASS cssRule.type is 1
PASS declaration.length is 0
PASS declaration.getPropertyValue('mix-blend-mode') is ""


Trailing comma : overlay,
PASS cssRule.type is 1
PASS declaration.length is 0
PASS declaration.getPropertyValue('mix-blend-mode') is ""
PASS successfullyParsed is true

TEST COMPLETE
